requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
wakeonlan==3.1.0
paramiko==3.4.0
pydantic==2.5.3
